Bugfix: wrong URL while setting up tests
