+++ e2e-tests/attributeSelector.spec.ts
@@ -1,33 +1,7 @@
 import { test, expect } from '@playwright/test';
-import mockData from '../playwright/mock-data/simpsons/simpsons_data.json';
-import mockAnnotations from '../playwright/mock-data/simpsons/simpsons_annotations.json';
-import mockAltText from '../playwright/mock-data/simpsons/simpsons_alttxt.json';
+import { beforeTest } from './common';
 
-test.beforeEach(async ({ page }) => {
-  await page.route('*/**/api/**', async (route) => {
-    const url = route.request().url();
-    let json;
-
-    if (url) {
-      if (url.includes('workspaces/Upset%20Examples/tables/simpsons/rows/?limit=9007199254740991')) {
-        json = mockData;
-        await route.fulfill({ json });
-      } else if (url.includes('workspaces/Upset%20Examples/tables/simpsons/annotations/')) {
-        json = mockAnnotations;
-        await route.fulfill({ json });
-      } else if (url.includes('alttxt')) {
-        json = mockAltText;
-        await route.fulfill({ json });
-      } else if (url.includes('workspaces/Upset%20Examples/sessions/table/193/state/')) {
-        await route.fulfill({ status: 200 });
-      } else {
-        await route.continue();
-      }
-    } else {
-      await route.abort();
-    }
-  });
-});
+test.beforeEach(beforeTest);
 
 test('Attribute Dropdown', async ({ page }) => {
   await page.goto('http://localhost:3000/?workspace=Upset+Examples&table=simpsons&sessionId=193');
